improvement: coloring of outdated packages
previously we had light red with background on hover being darker blue-ish. That hurts. So now it's a darker shade of red - somewhat in line with the darker blue which is used as the main color (though not kuler-proof yet) and has a light red background on hovering.

Also needed to restructure the element where coloring is done, since the outdated info occurs one div further down as the element that was used for coloring before

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>