add hover animation to version fields separate from package field
i.e. when hovering over a package summary, the color is slightly lighter than when hovering over its version elements. the versions cannot be styled separately as of right now because the :hover selector in css can't style the parent element. javascript is the proper way to solve this, but slightly overkill for this first version. it may be adjusted later on, to only style the version fields when it's hovered over and not the parent field...

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>