improvement: restyled node and package selection items to support subselection
this is important for version selection of packages: hover over child elements to only animate gives the freedom to only animate the child (as done here) or both child and parent. also adjusted the node scss to since package and node list both use the same styling, but they look and behave slightly differently: for node list the user always selects only the node element as a whole, while for packages the user either selects the package (ie. all versions of that package) or a number of versions of a package he is interested in

Signed-off-by: Dominik Richter <dominik.richter@gmail.com>